refactor: Split execution logic from TaskManager files
