+++ src/Craned/TaskManager.cpp
@@ -1385,266 +1385,4 @@ bool TaskManager::MigrateProcToCgroupOfTask(pid_t pid, task_id_t task_id) {
   return cg_ptr->MigrateProcIn(pid);
 }
 
-CraneErr ProcessInstance::Spawn(const task_id_t task_id,
-                                const PasswordEntry& pwd_entry,
-                                util::Cgroup* cgroup, EnvVector task_envs) {
-  using google::protobuf::io::FileInputStream;
-  using google::protobuf::io::FileOutputStream;
-  using google::protobuf::util::ParseDelimitedFromZeroCopyStream;
-  using google::protobuf::util::SerializeDelimitedToZeroCopyStream;
-
-  using crane::grpc::subprocess::CanStartMessage;
-  using crane::grpc::subprocess::ChildProcessReady;
-
-  int socket_pair[2];
-
-  if (socketpair(AF_UNIX, SOCK_STREAM, 0, socket_pair) != 0) {
-    CRANE_ERROR("Failed to create socket pair: {}", strerror(errno));
-    return CraneErr::kSystemErr;
-  }
-
-  // save the current uid/gid
-  std::pair<uid_t, gid_t> saved_priv{getuid(), getgid()};
-
-  int rc = setegid(pwd_entry.Gid());
-  if (rc == -1) {
-    CRANE_ERROR("error: setegid. {}", strerror(errno));
-    return CraneErr::kSystemErr;
-  }
-  __gid_t gid_a[1] = {pwd_entry.Gid()};
-  setgroups(1, gid_a);
-  rc = seteuid(pwd_entry.Uid());
-  if (rc == -1) {
-    CRANE_ERROR("error: seteuid. {}", strerror(errno));
-    return CraneErr::kSystemErr;
-  }
-
-  pid_t child_pid = fork();
-  if (child_pid > 0) {  // Parent proc
-    close(socket_pair[1]);
-    int fd = socket_pair[0];
-    bool ok;
-    CraneErr err;
-
-    setegid(saved_priv.second);
-    seteuid(saved_priv.first);
-    setgroups(0, nullptr);
-
-    FileInputStream istream(fd);
-    FileOutputStream ostream(fd);
-    CanStartMessage msg;
-    ChildProcessReady child_process_ready;
-
-    CRANE_DEBUG("Subprocess was created for task #{} pid: {}", task_id,
-                child_pid);
-
-    SetPid(child_pid);
-
-    // Add event for stdout/stderr of the new subprocess
-    // struct bufferevent* ev_buf_event;
-    // ev_buf_event =
-    //     bufferevent_socket_new(m_ev_base_, fd, BEV_OPT_CLOSE_ON_FREE);
-    // if (!ev_buf_event) {
-    //   CRANE_ERROR(
-    //       "Error constructing bufferevent for the subprocess of task #!",
-    //       instance->task.task_id());
-    //   err = CraneErr::kLibEventError;
-    //   goto AskChildToSuicide;
-    // }
-    // bufferevent_setcb(ev_buf_event, EvSubprocessReadCb_, nullptr, nullptr,
-    //                   (void*)process.get());
-    // bufferevent_enable(ev_buf_event, EV_READ);
-    // bufferevent_disable(ev_buf_event, EV_WRITE);
-    // process->SetEvBufEvent(ev_buf_event);
-
-    // Migrate the new subprocess to newly created cgroup
-    if (!cgroup->MigrateProcIn(GetPid())) {
-      CRANE_ERROR(
-          "Terminate the subprocess of task #{} due to failure of cgroup "
-          "migration.",
-          task_id);
-
-      err = CraneErr::kCgroupError;
-      goto AskChildToSuicide;
-    }
-
-    CRANE_TRACE("New task #{} is ready. Asking subprocess to execv...",
-                task_id);
-
-    // Tell subprocess that the parent process is ready. Then the
-    // subprocess should continue to exec().
-    msg.set_ok(true);
-    ok = SerializeDelimitedToZeroCopyStream(msg, &ostream);
-    ok &= ostream.Flush();
-    if (!ok) {
-      CRANE_ERROR("Failed to send ok=true to subprocess {} for task #{}",
-                  child_pid, task_id);
-      close(fd);
-      return CraneErr::kProtobufError;
-    }
-
-    ParseDelimitedFromZeroCopyStream(&child_process_ready, &istream, nullptr);
-    if (!msg.ok()) {
-      CRANE_ERROR("Failed to read protobuf from subprocess {} of task #{}",
-                  child_pid, task_id);
-      close(fd);
-      return CraneErr::kProtobufError;
-    }
-
-    close(fd);
-    return CraneErr::kOk;
-
-  AskChildToSuicide:
-    msg.set_ok(false);
-
-    ok = SerializeDelimitedToZeroCopyStream(msg, &ostream);
-    close(fd);
-    if (!ok) {
-      CRANE_ERROR("Failed to ask subprocess {} to suicide for task #{}",
-                  child_pid, task_id);
-      return CraneErr::kProtobufError;
-    }
-    return err;
-  } else {  // Child proc
-    rc = chdir(m_cwd_.c_str());
-    if (rc == -1) {
-      CRANE_ERROR("[Child Process] Error: chdir to {}. {}", m_cwd_.c_str(),
-                  strerror(errno));
-      std::abort();
-    }
-
-    setreuid(pwd_entry.Uid(), pwd_entry.Uid());
-    setregid(pwd_entry.Gid(), pwd_entry.Gid());
-
-    // Set pgid to the pid of task root process.
-    setpgid(0, 0);
-
-    close(socket_pair[0]);
-    int fd = socket_pair[1];
-
-    FileInputStream istream(fd);
-    FileOutputStream ostream(fd);
-    CanStartMessage msg;
-    ChildProcessReady child_process_ready;
-    bool ok;
-
-    ParseDelimitedFromZeroCopyStream(&msg, &istream, nullptr);
-    if (!msg.ok()) std::abort();
-
-    auto out_file_path = m_batch_meta_.parsed_output_file_pattern;
-    int out_fd = open(out_file_path.c_str(), O_RDWR | O_CREAT | O_APPEND, 0644);
-    if (out_fd == -1) {
-      CRANE_ERROR("[Child Process] Error: open {}. {}", out_file_path,
-                  strerror(errno));
-      std::abort();
-    }
-
-    dup2(out_fd, 1);  // stdout -> output file
-    dup2(out_fd, 2);  // stderr -> output file
-    close(out_fd);
-
-    child_process_ready.set_ok(true);
-    ok = SerializeDelimitedToZeroCopyStream(child_process_ready, &ostream);
-    ok &= ostream.Flush();
-    if (!ok) {
-      CRANE_ERROR("[Child Process] Error: Failed to flush.");
-      std::abort();
-    }
-
-    close(fd);
-
-    // If these file descriptors are not closed, a program like mpirun may
-    // keep waiting for the input from stdin or other fds and will never end.
-    close(0);  // close stdin
-    util::CloseFdFrom(3);
-
-    // Set environment variables related to task info (begin with `CRANE`).
-    EnvVector env_vec{std::move(task_envs)};
-
-    // Since we want to reinitialize the environment variables of the user,
-    // we are actually performing two steps: login -> start shell.
-    // Shell starting is done by calling "bash --login".
-    //
-    // During shell starting step, /etc/profile, ~/.bash_profile, ... are
-    // loaded.
-    //
-    // During login step, "HOME" and "SHELL" are set.
-    // Here we are just mimicking the login module.
-    //
-    // Slurm uses `su <username> -c /usr/bin/env` to retrieve
-    // all the environment variables.
-    // We use a more tidy way.
-
-    // FIXME: Check instance->task.get_user_env():
-    env_vec.emplace_back("HOME", pwd_entry.HomeDir());
-    env_vec.emplace_back("SHELL", pwd_entry.Shell());
-
-    if (clearenv()) {
-      fmt::print("clearenv() failed!\n");
-    }
-
-    for (const auto& [name, value] : env_vec) {
-      if (setenv(name.c_str(), value.c_str(), 1)) {
-        fmt::print("setenv for {}={} failed!\n", name, value);
-      }
-    }
-
-    // Prepare the command line arguments.
-    std::vector<const char*> argv;
-
-    // Argv[0] is the program name which can be anything.
-    argv.emplace_back("CraneScript");
-
-    // FIXME: Check instance->task.get_user_env():
-    argv.emplace_back("--login");
-
-    argv.emplace_back(m_executive_path_.c_str());
-    for (auto&& arg : m_arguments_) {
-      argv.push_back(arg.c_str());
-    }
-    argv.push_back(nullptr);
-
-    execv("/bin/bash", const_cast<char* const*>(argv.data()));
-
-    // Error occurred since execv returned. At this point, errno is set.
-    // Ctld use SIGABRT to inform the client of this failure.
-    fmt::print(stderr, "[Craned Subprocess Error] Failed to execv. Error: {}\n",
-               strerror(errno));
-    // Todo: See https://tldp.org/LDP/abs/html/exitcodes.html, return standard
-    //  exit codes
-    abort();
-  }
-}
-
-CraneErr ProcessInstance::Kill(int signum) {
-  // Todo: Add timer which sends SIGTERM for those tasks who
-  //  will not quit when receiving SIGINT.
-  if (!m_pid_) {
-    // Send the signal to the whole process group.
-    int err = kill(-m_pid_, signum);
-
-    if (err == 0)
-      return CraneErr::kOk;
-    else {
-      CRANE_TRACE("kill pid {} failed. error: {}", m_pid_, strerror(errno));
-      return CraneErr::kGenericFailure;
-    }
-  }
-
-  return CraneErr::kNonExistent;
-}
-
-CraneErr ContainerInstance::Spawn(task_id_t task_id,
-                                  const PasswordEntry& pwd_entry,
-                                  util::Cgroup* cgroup,
-                                  Craned::TaskExecutor::EnvVector task_envs) {
-  // TODO: Implement this
-  return CraneErr::kOk;
-}
-
-CraneErr ContainerInstance::Kill(int signum) {
-  // TODO: Implement this
-  return CraneErr::kOk;
-}
-
 }  // namespace Craned
